projeto_etiquetas/requirements.txt
pandas==2.0.3
openpyxl==3.1.2
reportlab==4.0.4
tkinter
sqlite3
Pillow==10.0.0
